requirements.txt
streamlit
pandas
yfinance
matplotlib
openai
fredapi
python-dotenv
request
datetime
